third_party/blink/web_tests/storage/domstorage/clear-expected.txt

Test basic dom storage .clear() functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing sessionStorage
storage.clear()
PASS storage.length is 0
storage['FOO'] = 'MyFOO'
storage['BAR'] = 'MyBar'
PASS storage.length is 2
PASS storage['FOO'] is "MyFOO"
PASS storage['BAR'] is "MyBar"
storage.clear()
PASS storage.length is 0
PASS storage['FOO'] is undefined
PASS storage['BAR'] is undefined




Testing localStorage
storage.clear()
PASS storage.length is 0
storage['FOO'] = 'MyFOO'
storage['BAR'] = 'MyBar'
PASS storage.length is 2
PASS storage['FOO'] is "MyFOO"
PASS storage['BAR'] is "MyBar"
storage.clear()
PASS storage.length is 0
PASS storage['FOO'] is undefined
PASS storage['BAR'] is undefined
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
